requirements.txt
SQLAlchemy==1.4.42
uvicorn==0.23.2
Jinja2==3.1.2
mysql==0.0.3
fastapi==0.101.0
passlib==1.7.4
python-jose==3.3.0
pydantic==2.4.2
PyJWT==2.8.0
